Temporarily fixing fieldname issues
Noms does not support '-' in field names, so
I'm replacing them with '_'.

We have to undo this at some point if we want reversibility,
but this works at least to get the data in there.

I was hoping this was the cause of the other issues, but it's still
failing with different errors.

Signed-off-by: Luke Kanies <luke@madstop.com>